Clarify note about item_id for shipping addresses
Based on user feedback, it was not easy for the user to determine what to use as the `item_id` in a Create Refund Quote request should be for a product with `"item_type": "SHIPPING"`. The FAQ section of this page calls out that it is the order address ID as does the example request body in the reference material. There is a note on this page that calls out the source for `item_id` for a product with `"item_type": "PRODUCT"`, but that note does not mention shipping addresses. This update adds a callout about the Order Shipping Addresses API to that note.