(external): [update] Catalog Brands, update meta_keywords descriptions and examples (#50)
While working through the API I noticed the description and examples for
meta_kaywords don't reflect what is received back from the actual
request. This hopefully should fix that.

- Sorry for accidentally deleting the last one.

<!-- Ticket number or summary of work -->
# Update meta_keywords Descriptions and Examples


## What changed?
<!-- Provide a bulleted list in the present tense -->
* Changed description from "Comma-separated list of meta keywords to
include in the HTML." to "An array of meta keywords to include in the
HTML."
* Changed examples to be an array of keywords and not a comma-separated
string

## Release notes draft
Update Brand's V3 documentation to correctly represent what is returned
in the response.